feat: add career link to user dropdown
fix: addressing linter errors

chore: update tests snapshot

chore: add career link to user dropdown to header variant

chore: remove header and rename header variant to be header

chore: only enterprise user should see the link

chore: only non-enterprise student should see the item

feat: upgraded to node v18, added .nvmrc and updated workflows (#151)
