+++ .env
@@ -39,3 +39,4 @@ HOTJAR_DEBUG=''
 ACCOUNT_SETTINGS_URL=''
 ACCOUNT_PROFILE_URL=''
 ENABLE_NOTICES=''
+CAREER_LINK_URL=''
