+++ .env.development
@@ -46,3 +46,4 @@ HOTJAR_DEBUG=''
 ACCOUNT_SETTINGS_URL='http://localhost:1997'
 ACCOUNT_PROFILE_URL='http://localhost:1995'
 ENABLE_NOTICES=''
+CAREER_LINK_URL=''
