+++ .env.test
@@ -45,3 +45,4 @@ HOTJAR_DEBUG=''
 ACCOUNT_SETTINGS_URL='http://account-settings-url.test'
 ACCOUNT_PROFILE_URL='http://account-profile-url.test'
 ENABLE_NOTICES=''
+CAREER_LINK_URL=''
