+++ package.json
@@ -77,7 +77,7 @@
     "whatwg-fetch": "^3.6.2"
   },
   "devDependencies": {
-    "@edx/frontend-build": "11.0.1",
+    "@edx/frontend-build": "12.8.27",
     "@edx/reactifex": "^2.1.1",
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^12.1.0",
@@ -91,6 +91,6 @@
     "react-dev-utils": "^11.0.4",
     "react-test-renderer": "^16.14.0",
     "redux-mock-store": "^1.5.4",
-    "semantic-release": "^17.4.5"
+    "semantic-release": "^20.1.3"
   }
 }
