+++ src/containers/LearnerDashboardHeader/CollapsedHeader/index.test.jsx
@@ -2,14 +2,14 @@ import { shallow } from 'enzyme';
 
 import CollapsedHeader from '.';
 
-import { useLearnerDashboardHeaderVariantData, useIsCollapsed } from '../hooks';
+import { useLearnerDashboardHeaderData, useIsCollapsed } from '../hooks';
 
 jest.mock('../BrandLogo', () => jest.fn(() => 'BrandLogo'));
 jest.mock('./CollapseMenuBody', () => jest.fn(() => 'CollapseMenuBody'));
 
 jest.mock('../hooks', () => ({
   useIsCollapsed: jest.fn(() => true),
-  useLearnerDashboardHeaderVariantData: jest.fn(() => ({
+  useLearnerDashboardHeaderData: jest.fn(() => ({
     isOpen: false,
     toggleIsOpen: jest.fn().mockName('toggleIsOpen'),
   })),
@@ -28,7 +28,7 @@ describe('CollapsedHeader', () => {
   });
 
   it('renders with isOpen true', () => {
-    useLearnerDashboardHeaderVariantData.mockReturnValueOnce({
+    useLearnerDashboardHeaderData.mockReturnValueOnce({
       isOpen: true,
       toggleIsOpen: jest.fn().mockName('toggleIsOpen'),
     });
