+++ src/containers/LearnerDashboardHeader/hooks.js
@@ -1,11 +1,18 @@
 import React from 'react';
 import { useWindowSize, breakpoints } from '@edx/paragon';
 import track from 'tracking';
+import { StrictDict } from 'utils';
 import { linkNames } from 'tracking/constants';
+
+import * as module from './hooks';
+
+export const state = StrictDict({
+  isOpen: (val) => React.useState(val), // eslint-disable-line
+});
 
 export const useIsCollapsed = () => {
   const { width } = useWindowSize();
-  const isCollapsed = React.useMemo(() => (width <= breakpoints.large.maxWidth), [width]);
+  const isCollapsed = React.useMemo(() => (width <= breakpoints.large.minWidth), [width]);
   return isCollapsed;
 };
 
@@ -17,8 +24,19 @@ export const findCoursesNavDropdownClicked = (href) => track.findCourses.findCou
   linkName: linkNames.learnerHomeNavDropdownExplore,
 });
 
+export const useLearnerDashboardHeaderData = () => {
+  const [isOpen, setIsOpen] = module.state.isOpen(false);
+  const toggleIsOpen = () => setIsOpen(!isOpen);
+
+  return {
+    isOpen,
+    toggleIsOpen,
+  };
+};
+
 export default {
   useIsCollapsed,
   findCoursesNavClicked,
   findCoursesNavDropdownClicked,
+  useLearnerDashboardHeaderData,
 };
