+++ src/containers/LearnerDashboardHeader/hooks.test.js
@@ -1,7 +1,19 @@
 import { useWindowSize, breakpoints } from '@edx/paragon';
 import track from 'tracking';
 import { linkNames } from 'tracking/constants';
-import { useIsCollapsed, findCoursesNavClicked, findCoursesNavDropdownClicked } from './hooks';
+
+import { MockUseState } from 'testUtils';
+
+import * as hooks from './hooks';
+
+const state = new MockUseState(hooks);
+
+const {
+  useIsCollapsed,
+  findCoursesNavClicked,
+  findCoursesNavDropdownClicked,
+  useLearnerDashboardHeaderData,
+} = hooks;
 
 jest.mock('tracking', () => ({
   findCourses: {
@@ -12,13 +24,17 @@ jest.mock('tracking', () => ({
 const url = 'http://example.com';
 
 describe('LearnerDashboardHeader hooks', () => {
+  describe('state values', () => {
+    state.testGetter(state.keys.isOpen);
+  });
+
   describe('useIsCollapsed', () => {
     test('large screen is not collapsed', () => {
-      useWindowSize.mockReturnValueOnce({ width: breakpoints.large.maxWidth + 1 });
+      useWindowSize.mockReturnValueOnce({ width: breakpoints.large.minWidth + 1 });
       expect(useIsCollapsed()).toEqual(false);
     });
     test('small screen is collapsed', () => {
-      useWindowSize.mockReturnValueOnce({ width: breakpoints.large.maxWidth - 1 });
+      useWindowSize.mockReturnValueOnce({ width: breakpoints.large.minWidth - 1 });
       expect(useIsCollapsed()).toEqual(true);
     });
   });
@@ -38,6 +54,16 @@ describe('LearnerDashboardHeader hooks', () => {
       expect(track.findCourses.findCoursesClicked).toHaveBeenCalledWith(url, {
         linkName: linkNames.learnerHomeNavDropdownExplore,
       });
+    });
+  });
+
+  describe('useLearnerDashboardHeaderData', () => {
+    test('default state', () => {
+      state.mock();
+      const out = useLearnerDashboardHeaderData();
+      state.expectInitializedWith(state.keys.isOpen, false);
+      out.toggleIsOpen();
+      expect(state.values.isOpen).toEqual(true);
     });
   });
 });
