+++ src/containers/LearnerDashboardHeader/index.scss
@@ -1,12 +1,38 @@
-.greetings-slash-container-small {
-  height: 4px;
-  width: 40px;
-  transform-origin: center;
-  transform: rotate(-70deg);
+.dropdown-menu-collapse {
+  width: 100vw;
+  position: absolute;
+  left: 0;
 }
-.greetings-slash-container-large {
-  height: 8px;
-  width: 120px;
-  transform-origin: center;
-  transform: rotate(-70deg);
+
+.learner-variant-header {
+  a {
+    // needed to make the link not resize the header
+    border-bottom: 2px solid transparent;
+  }
+  .course-link {
+    border-bottom: 2px solid !important;
+  }
+
+  .course-link:hover {
+    border-bottom: inherit !important;
+  }
+}
+
+.nav-small-menu {
+  > * {
+    justify-content: flex-start !important;
+
+    border-radius: 0 !important;
+    border-top: 1px solid #ddd !important;
+
+    &::after {
+      content: '\00BB';
+      padding-left: 10px;
+    }
+  }
+}
+
+.logo {
+  // copy from legacy dashboard
+  height: 40px;
 }
