+++ src/containers/LearnerDashboardHeader/messages.js
@@ -2,70 +2,79 @@ import { defineMessages } from '@edx/frontend-platform/i18n';
 
 const messages = defineMessages({
   dashboard: {
-    id: 'leanerDashboard.menu.dashboard.label',
+    id: 'learnerVariantDashboard.menu.dashboard.label',
     defaultMessage: 'Dashboard',
     description: 'The text for the user menu Dashboard navigation link.',
   },
   help: {
-    id: 'leanerDashboard.help.label',
+    id: 'learnerVariantDashboard.help.label',
     defaultMessage: 'Help',
     description: 'The text for the link to the Help Center',
   },
   profile: {
-    id: 'leanerDashboard.menu.profile.label',
+    id: 'learnerVariantDashboard.menu.profile.label',
     defaultMessage: 'Profile',
     description: 'The text for the user menu Profile navigation link.',
   },
   viewPrograms: {
-    id: 'leanerDashboard.menu.viewPrograms.label',
+    id: 'learnerVariantDashboard.menu.viewPrograms.label',
     defaultMessage: 'View Programs',
     description: 'The text for the user menu View Programs navigation link.',
   },
   account: {
-    id: 'leanerDashboard.menu.account.label',
+    id: 'learnerVariantDashboard.menu.account.label',
     defaultMessage: 'Account',
     description: 'The text for the user menu Account navigation link.',
   },
   orderHistory: {
-    id: 'leanerDashboard.menu.orderHistory.label',
+    id: 'learnerVariantDashboard.menu.orderHistory.label',
     defaultMessage: 'Order History',
     description: 'The text for the user menu Order History navigation link.',
   },
   signOut: {
-    id: 'leanerDashboard.menu.signOut.label',
+    id: 'learnerVariantDashboard.menu.signOut.label',
     defaultMessage: 'Sign Out',
     description: 'The label for the user menu Sign Out action.',
   },
-
-  goodMorning: {
-    id: 'greeting.morning',
-    defaultMessage: 'Good Morning!',
-    description: 'Good Morning',
+  course: {
+    id: 'learnerVariantDashboard.course',
+    defaultMessage: 'Courses',
+    description: 'Header link for switching to dashboard page.',
   },
-  goodAfternoon: {
-    id: 'greeting.afternoon',
-    defaultMessage: 'Good Afternoon!',
-    description: 'Good Afternoon',
-  },
-  goodEvening: {
-    id: 'greeting.evening',
-    defaultMessage: 'Good Evening!',
-    description: 'Good Evening',
-  },
-  switchToProgram: {
-    id: 'leanerDashboard.switchToProgram',
-    defaultMessage: 'Switch to Programs',
+  program: {
+    id: 'learnerVariantDashboard.program',
+    defaultMessage: 'Programs',
     description: 'Header link for switching to program page.',
   },
-  exploreCourses: {
-    id: 'leanerDashboard.exploreCourses',
-    defaultMessage: 'Explore courses',
-    description: 'Header link for switching to course page.',
+  discoverNew: {
+    id: 'learnerVariantDashboard.discoverNew',
+    defaultMessage: 'Discover New',
+    description: 'Header link for switching to discover page.',
   },
-  courseSearchAlt: {
-    id: 'leanerDashboard.courseSearchAlt',
-    defaultMessage: 'Course search',
-    description: 'Alt-text for course search icon button',
+  logoAltText: {
+    id: 'learnerVariantDashboard.logoAltText',
+    defaultMessage: 'edX, Inc. Dashboard',
+    description: 'Alt text for the edX logo.',
+  },
+  collapseMenuOpenAltText: {
+    id: 'learnerVariantDashboard.collapseMenuOpenAltText',
+    defaultMessage: 'Menu',
+    description: 'Alt text for the collapse menu icon when the menu is open.',
+  },
+  collapseMenuClosedAltText: {
+    id: 'learnerVariantDashboard.collapseMenuClosedAltText',
+    defaultMessage: 'Close',
+    description: 'Alt text for the collapse menu icon when the menu is closed.',
+  },
+  career: {
+    id: 'leanerDashboard.menu.career.label',
+    defaultMessage: 'Career',
+    description: 'The text for the user menu Career navigation link.',
+  },
+  newAlert: {
+    id: 'header.menu.new.label',
+    defaultMessage: 'New',
+    description: 'The text announcing that an item in the user menu is New',
   },
 });
