+++ src/data/redux/hooks/app.js
@@ -1,3 +1,4 @@
+/* eslint-disable react-hooks/rules-of-hooks */
 import { useSelector, useDispatch } from 'react-redux';
 
 import * as redux from 'data/redux';
